--- README.md
@@ -1,9 +1,0 @@
-# **WorkMate**
-
-## > *"The Modern Day Golden Pages"*
-
-## **Our Goals**
-1. Connect the Consumer and Service Provider in a safe and secure fashion
-2. Provide an easy-to-use interface for both parties
-3. Useful functionality such as geo-location, reviewal systems and messaging
-4. To finish it before the end of term lol
