+++ app/build.gradle
@@ -1,7 +1,6 @@
 plugins {
     id 'com.android.application'
     id 'com.google.gms.google-services'
-    id 'com.google.firebase.crashlytics'
 }
 
 android {
@@ -36,20 +35,7 @@ dependencies {
     implementation 'com.google.android.material:material:1.3.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
     implementation 'androidx.wear:wear:1.1.0'
-    implementation 'com.google.firebase:firebase-auth:20.0.3'
-    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-
-    implementation 'com.google.android.gms:play-services-auth:19.0.0'
-
-    //Added for messaging
-    implementation platform('com.google.firebase:firebase-bom:27.0.0')
-    implementation 'com.google.firebase:firebase-database'
-    implementation 'com.google.firebase:firebase-storage'
-    implementation 'com.google.firebase:firebase-analytics'
-    implementation 'com.firebaseui:firebase-ui-database:7.1.1'
-    implementation 'com.github.bumptech.glide:glide:4.12.0'
-    implementation 'com.google.firebase:firebase-inappmessaging-display'
-    implementation 'com.google.firebase:firebase-messaging'
+    implementation 'com.google.firebase:firebase-auth:20.0.2'
 
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.1'
     testImplementation 'org.testng:testng:6.9.6'
@@ -60,24 +46,4 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
     compileOnly 'com.google.android.wearable:wearable:2.8.1'
     implementation 'de.hdodenhof:circleimageview:3.1.0'
-
-    //Messaging
-    implementation 'com.github.bumptech.glide:glide:4.12.0'
-    implementation 'de.hdodenhof:circleimageview:3.1.0'
-    implementation platform('com.google.firebase:firebase-bom:26.8.0')
-    implementation 'com.google.firebase:firebase-analytics'
-    implementation 'com.google.firebase:firebase-config'
-    implementation 'com.google.firebase:firebase-crashlytics'
-    implementation 'com.google.firebase:firebase-database'
-    implementation 'com.google.firebase:firebase-inappmessaging-display'
-    implementation 'com.google.firebase:firebase-messaging'
-    implementation 'com.google.firebase:firebase-storage'
-    implementation 'com.firebaseui:firebase-ui-database:7.1.1'
-    implementation 'androidx.arch.core:core-runtime:2.1.0'
-    implementation 'androidx.arch.core:core-common:2.1.0'
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
-    androidTestImplementation 'androidx.test:runner:1.3.0'
-    androidTestImplementation 'androidx.test:rules:1.3.0'
 }
