+++ app/src/main/java/com/example/workmate/ClientModel.java
@@ -1,4 +1,4 @@
-package com.example.workmate.Models;
+package com.example.workmate;
 
 
 //this class defines the data type client, lists its data, and provides constructors, a print function, and getters and setters
@@ -12,10 +12,6 @@ public class ClientModel {
     private String ClientPhone;
     private String ClientEmail;
     private String ClientAddr;
-    private String ClientLatitude;
-    private String ClientLongitude;
-
-
 
     //constructors
     public ClientModel(int clientId, String clientFname, String clientLname, String clientPhone, String clientEmail,  String clientAddr) {
@@ -93,19 +89,5 @@ public class ClientModel {
 
     public void setClientAddr(String clientAddr) {
         ClientAddr = clientAddr;
-    }
-
-    public String getClientLatitude() {return ClientLatitude; }
-
-    public void setClientLatitude(String clientLatitude) {
-        //nb this needs to be taken from the google api BEFORE ENTERING IN SQL
-        ClientLatitude = clientLatitude;
-    }
-
-    public String getClientLongitude() {return ClientLongitude; }
-
-    public void setClientLongitude(String clientLongitude) {
-        //nb this needs to be taken from the google api BEFORE ENTERING IN SQL
-        ClientLongitude = clientLongitude;
     }
 }
