--- app/src/main/java/com/example/workmate/CodelabPreferences.java
@@ -1,7 +1,0 @@
-package com.example.workmate;
-
-public class CodelabPreferences {
-
-    public static final String FRIENDLY_MSG_LENGTH = "friendly_msg_length";
-
-}
