+++ app/src/main/java/com/example/workmate/HomeFragment.java
@@ -1,4 +1,4 @@
-package com.example.workmate.Fragments;
+package com.example.workmate;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -8,8 +8,6 @@ import android.widget.Button;
 
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-
-import com.example.workmate.R;
 
 public class HomeFragment extends Fragment {
