+++ app/src/main/java/com/example/workmate/LoginFragment.java
@@ -1,0 +1,25 @@
+package com.example.workmate;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+public class LoginFragment extends Fragment {
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState){
+
+
+
+        return inflater.inflate(R.layout.fragment_login,  container, false);
+    }
+
+
+
+
+}
