+++ app/src/main/java/com/example/workmate/MessagesFragment.java
@@ -1,0 +1,17 @@
+package com.example.workmate;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+public class MessagesFragment extends Fragment {
+
+    @Nullable
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState){
+        return inflater.inflate(R.layout.fragment_message,  container, false);
+    }
+}
