--- app/src/main/java/com/example/workmate/MyFirebaseMessagingService.java
@@ -1,19 +1,0 @@
-package com.example.workmate;
-
-import android.util.Log;
-
-import com.google.firebase.messaging.FirebaseMessaging;
-import com.google.firebase.messaging.FirebaseMessagingService;
-import com.google.firebase.messaging.RemoteMessage;
-
-public class MyFirebaseMessagingService {
-
-    private static final String TAG = "MyFMService";
-
-    public void onMessageReceived(RemoteMessage remoteMessage) {
-        // Handle data payload of FCM messages.
-        Log.d(TAG, "FCM Message Id: " + remoteMessage.getMessageId());
-        Log.d(TAG, "FCM Notification Message: " + remoteMessage.getNotification());
-        Log.d(TAG, "FCM Data Message: " + remoteMessage.getData());
-    }
-}
