+++ app/src/main/java/com/example/workmate/RVAdapter.java
@@ -1,4 +1,4 @@
-package com.example.workmate.Adapters;
+package com.example.workmate;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -8,9 +8,6 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
-
-import com.example.workmate.R;
-import com.example.workmate.Models.SupplierModel;
 
 import java.util.ArrayList;
