--- app/src/main/res/drawable/rounded_corners.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="#1c9403" />
-    <!--border color and border width-->
-    <stroke
-        android:width="7dp"
-        android:color="@color/white" />
-    <!--border radius in dp-->
-    <corners android:radius="40dp" />
-</shape>
