+++ app/src/main/res/layout/activity_client_reg.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".Activities.MainActivity">
+    tools:context=".MainActivity">
 
 
     <include
@@ -17,7 +17,7 @@
     <ScrollView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".Activities.ClientRegActivity">
+        tools:context=".ClientRegActivity">
 
         <LinearLayout
             android:layout_width="match_parent"
@@ -184,7 +184,7 @@
                 android:text="@string/toc" />
 
             <Button
-                android:id="@+id/client_reg"
+                android:id="@+id/button"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:text="@string/title_activity_register" />
