+++ app/src/main/res/layout/activity_service_reg.xml
@@ -15,7 +15,7 @@
     <ScrollView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".Activities.ClientRegActivity">
+        tools:context=".ClientRegActivity">
 
         <LinearLayout
             android:layout_width="match_parent"
@@ -187,7 +187,7 @@
                 android:text="@string/toc" />
 
             <Button
-                android:id="@+id/service_reg"
+                android:id="@+id/button"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:text="@string/title_activity_register" />
