+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".Activities.MainActivity">
+    tools:context=".MainActivity">
 
     <TextView
         android:id="@+id/textView4"
