+++ app/src/main/res/layout/fragment_profile.xml
@@ -11,7 +11,7 @@
         android:layout_height="match_parent"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        tools:context=".Activities.ClientRegActivity"
+        tools:context=".ClientRegActivity"
         tools:layout_editor_absoluteX="39dp">
 
         <LinearLayout
