+++ app/src/main/res/values-round/strings.xml
@@ -1,5 +1,3 @@
 <resources>
 
-
-    <string name="hello_world">Hello Round World!</string>
 </resources>
