+++ app/src/main/res/values/strings.xml
@@ -69,13 +69,6 @@
     <string name="view_suppliers">View Suppliers</string>
     <string name="viewdata">ViewData</string>
     <string name="search">Search&#8230;</string>
-    <string name="drawer_editprofile">Edit Profile</string>
-    <string name="title_activity_edit_profile">EditProfile</string>
-    <!--
-    This string is used for square devices and overridden by hello_world in
-    values-round/strings.xml for round devices.
-    -->
-    <string name="hello_world">Hello Square World!</string>
 
 
 </resources>
