--- insertValues.php
@@ -1,20 +1,0 @@
-<?php
-
-$link = mysqli_connect("localhost", "***", "***", "workmatedb");
-
-// Check connection
-if($link === false){
-    die("ERROR: Could not connect. " . mysqli_connect_error());
-}
-
-// Attempt insert query execution
-$sql = "INSERT INTO client_table (client_fname, client_lname, client_email, client_phone, client_addr) VALUES ('Peter', '',....);";
-if(mysqli_query($link, $sql)){
-    echo "Records inserted successfully.";
-} else{
-    echo "ERROR: Could not able to execute $sql. " . mysqli_error($link);
-}
-
-// Close connection
-mysqli_close($link);
-?>
