--- validate.php
@@ -1,10 +1,0 @@
-//this file cleans up the input
-<?php
-function validate($data) {
-  $data = trim($data);
-  $data = stripslashes($data);
-  $data = htmlspecialchars($data);
-
-  return $data;
-}
- ?>
